* make-etc / activate-configuration: allow /etc files to be installed   as regular files instead of symlinks to the store. * Sudo configuration, enabled through security.sudo.enable (on by   default).  The contents of the sudoers file is specified in   security.sudo.configFile.  The default sudoers file allows members   of the new "wheel" group to run any command.
svn path=/nixos/trunk/; revision=9138

+++ system/activate-configuration.sh
@@ -20,7 +20,15 @@ ln -s @etc@/etc $staticEtc
 for i in $(cd $staticEtc && find * -type l); do
     mkdir -p /etc/$(dirname $i)
     rm -f /etc/$i
-    ln -s $staticEtc/$i /etc/$i
+    if test -e "$staticEtc/$i.mode"; then
+        # Create a regular file in /etc.
+        cp $staticEtc/$i /etc/$i
+        chown root.root /etc/$i
+        chmod "$(cat "$staticEtc/$i.mode")" /etc/$i
+    else
+        # Create a symlink in /etc.
+        ln -s $staticEtc/$i /etc/$i
+    fi
 done
